foundation/orchestration/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langgraph
psycopg2-binary
langgraph-checkpoint-postgres
python-dotenv
httpx
